Refactoring inside userControllerTest package
